+++ pioreactor/types.py
@@ -83,6 +83,11 @@ PwmChannel = t.Literal["1", "2", "3", "4", "5"]
 PdAngle = t.Literal["45", "90", "135", "180"]
 PdAngleOrREF = t.Union[PdAngle, t.Literal["REF"]]
 
+# hardware level stuff
+AnalogValue = t.Union[int, float]
+Voltage = float
+
+AdcChannel = t.Literal[0, 1, 2, 3]
 
 # All GPIO pins below are BCM numbered
 GpioPin = t.Literal[