+++ pioreactor/utils/mock.py
@@ -5,7 +5,10 @@ from __future__ import annotations
 import random
 from typing import Any
 
+import pioreactor.types as pt
 from pioreactor.config import config
+from pioreactor.utils.adcs import _ADC
+from pioreactor.utils.dacs import _DAC
 from pioreactor.whoami import am_I_active_worker
 from pioreactor.whoami import is_testing_env
 
@@ -33,37 +36,19 @@ class MockI2C:
         return
 
 
-class MockAnalogIn:
+class Mock_ADC(_ADC):
     INIT_STATE = 0.01
     state = INIT_STATE
     _counter = 0.0
     OFFSET = 0.03
 
-    def __init__(self, ads, channel, **kwargs) -> None:
+    def __init__(self, *args, **kwargs) -> None:
 
         self.max_gr = 0.25 + 0.1 * random.random()
         self.scale_factor = 0.00035 + 0.00005 * random.random()
         self.lag = 2 * 60 * 60 - 1 * 60 * 60 * random.random()
-        self.channel = channel
-        self.am_i_REF = str(channel + 1) == config.get(
-            "od_config.photodiode_channel_reverse", "REF"
-        )
 
-    def growth_rate(self, duration_as_seconds: float) -> float:
-        if self.am_i_REF:
-            return 0
-
-        import numpy as np
-
-        return (
-            self.max_gr
-            / (1 + np.exp(-self.scale_factor * (duration_as_seconds - self.lag)))
-            * (1 - 1 / (1 + np.exp(-self.scale_factor * 2 * (duration_as_seconds - 3 * self.lag))))
-        )
-
-    @property
-    def voltage(self) -> float:
-
+    def read_from_channel(self, channel: pt.AdcChannel):
         from pioreactor.utils import local_intermittent_storage
         import random
         import numpy as np
@@ -74,11 +59,13 @@ class MockAnalogIn:
         if not is_ir_on:
             return self.OFFSET
 
-        if self.am_i_REF:
+        am_i_REF = str(channel + 1) == config.get("od_config.photodiode_channel_reverse", "REF")
+
+        if am_i_REF:
             return (0.1 + random.normalvariate(0, sigma=0.001)) / 2**10 * 40 + self.OFFSET
         else:
             self.gr = self.growth_rate(
-                self._counter / config.getfloat("od_config", "samples_per_second")
+                self._counter / config.getfloat("od_config", "samples_per_second"), am_i_REF
             )
             self.state *= np.exp(
                 self.gr
@@ -88,18 +75,34 @@ class MockAnalogIn:
                 / 25  # divide by 25 from oversampling_count
             )
             self._counter += 1.0 / 25.0  # divide by 25 from oversampling_count
-            return self.state + random.normalvariate(0, sigma=self.state * 0.01) + self.OFFSET
+            return self.from_voltage_to_raw(
+                self.state + random.normalvariate(0, sigma=self.state * 0.01) + self.OFFSET
+            )
 
-    @property
-    def value(self) -> int:
-        return round(self.voltage * 32767 / 4.096)
+    def growth_rate(self, duration_as_seconds: float, am_i_REF: bool) -> float:
+        if am_i_REF:
+            return 0
+
+        import numpy as np
+
+        return (
+            self.max_gr
+            / (1 + np.exp(-self.scale_factor * (duration_as_seconds - self.lag)))
+            * (1 - 1 / (1 + np.exp(-self.scale_factor * 2 * (duration_as_seconds - 3 * self.lag))))
+        )
+
+    def check_on_gain(self, *args, **kwargs) -> None:
+        pass
+
+    def from_voltage_to_raw(self, voltage) -> int:
+        return round(voltage * 32767 / 4.096)
+
+    def from_raw_to_voltage(self, raw) -> float:
+        return 4.096 * raw / 32767
 
 
-class MockDAC43608:
+class Mock_DAC(_DAC):
 
-    _DEVICE_CONFIG = 1
-    _STATUS = 2
-    _BRDCAST = 3
     A = 8
     B = 9
     C = 10
@@ -109,19 +112,15 @@ class MockDAC43608:
     G = 14
     H = 15
 
-    def __init__(self, *args, **kwargs) -> None:
-        pass
-
-    def set_intensity_to(self, channel: str, intensity: float) -> None:
-        assert 0 <= intensity <= 1, "intensity should be between 0 and 1"
+    def set_intensity_to(self, channel: int, intensity: float) -> None:
+        assert 0.0 <= intensity <= 100.0, "intensity should be between 0 and 100"
         assert channel in list(range(8, 16)), "register should be in 8 to 15"
-        # TODO: this should update MQTT too
         return
 
-    def power_up(*args) -> None:
+    def power_up(self, *args) -> None:
         pass
 
-    def power_down(*args) -> None:
+    def power_down(self, *args) -> None:
         pass